html/modules/dynamicdata/xarutil.php: 	#  BitKeeper file /home/mikespub/xaraya/html/modules/dynamicdata/xarutil.php html/modules/dynamicdata/xarutilapi.php: 	#  BitKeeper file /home/mikespub/xaraya/html/modules/dynamicdata/xarutilapi.php html/modules/dynamicdata/xartemplates/util-static.xd: 	#  BitKeeper file /home/mikespub/xaraya/html/modules/dynamicdata/xartemplates/util-static.xd html/modules/dynamicdata/xartemplates/util-relations.xd: 	#  BitKeeper file /home/mikespub/xaraya/html/modules/dynamicdata/xartemplates/util-relations.xd html/modules/dynamicdata/sample.xml: 	#  move out utility functions html/modules/dynamicdata/xaradmin.php: 	#  move out utility functions html/modules/dynamicdata/xaradminapi.php: 	#  move out utility functions html/modules/dynamicdata/xaruserapi.php: 	#  move out utility functions html/modules/dynamicdata/xartemplates/admin-modify.xd: 	#  move out utility functions html/modules/dynamicdata/xartemplates/admin-modifyprop.xd: 	#  move out utility functions html/modules/dynamicdata/xartemplates/util-export.xd: 	#  move out utility functions html/modules/dynamicdata/xartemplates/util-import.xd: 	#  move out utility functions html/modules/dynamicdata/xartemplates/util-meta.xd: 	#  move out utility functions html/modules/dynamicdata/class/objects.php: 	#  move out utility functions html/modules/dynamicdata/class/properties.php: 	#  move out utility functions

+++ html/modules/dynamicdata/class/properties.php
@@ -229,7 +229,11 @@ class Dynamic_Property_Master
         // TODO: delete all the (dynamic ?) data for this property as well
         $object = new Dynamic_Object(array('objectid' => 2, // the Dynamic Properties = 2
                                            'itemid'   => $args['itemid']));
-        $object->getItem();
+        if (empty($object)) return;
+
+        $objectid = $object->getItem();
+        if (empty($objectid)) return;
+
         $objectid = $object->deleteItem();
         return $objectid;
     }
